Create good_subsequence
A sequence B is called good if B i m o d 2 ≠ B i − 1 m o d 2  for all 2 ≤ i ≤ ∣ B ∣. For example, [ 1 , 2 , 3 ]  and [ 4 ]  are good sequences, but [ 1 , 3 ]  is not. You are given an array A of length N. Find the length of the largest subsequence of A  that is good.